Updating google analytics ID

+++ www/js/app.js
@@ -25,15 +25,16 @@ angular.module('starter',
         'ngMaterial',
         'ngMessages',
         'angular-cache',
-        'angular-d3-word-cloud'
+        'angular-d3-word-cloud',
+        'ngFileUpload'
+        //'ui-iconpicker'
     ]
 )
-.run(function($ionicPlatform, $ionicHistory, $state, $rootScope, quantimodoService, Analytics, $ionicLoading) {
+.run(function($ionicPlatform, $ionicHistory, $state, $rootScope, quantimodoService, $http) {
+    window.developmentMode = window.location.href.indexOf("://localhost:") !== -1;
+    quantimodoService.getPrivateConfigs();
     quantimodoService.showBlackRingLoader();
-    if(appsManager.getUrlParameter('logout')){
-        localStorage.clear();
-        $rootScope.user = null;
-    }
+    if(appsManager.getUrlParameter('logout')){localStorage.clear(); $rootScope.user = null;}
     quantimodoService.setPlatformVariables();
     $ionicPlatform.ready(function() {
         //$ionicAnalytics.register();
@@ -182,16 +183,18 @@ angular.module('starter',
                 }
             }
         };
-        if(window.location.href.indexOf('ionic.quantimo') === -1 && window.location.href.indexOf('staging.quantimo') === -1 && window.location.href.indexOf('local.quantimo') === -1){
-            if(typeof analytics !== "undefined") {analytics.startTrackerWithId(config.appSettings.googleAnalyticsTrackingIds.ionic);}
-        }
         if (window.cordova && window.cordova.plugins && window.cordova.plugins.Keyboard) {
             cordova.plugins.Keyboard.hideKeyboardAccessoryBar(false); // Hide the accessory bar by default (remove this to show the accessory bar above the keyboard for form inputs
         }
         if (window.StatusBar) {StatusBar.styleDefault();} // org.apache.cordova.statusbar required
     });
-    $rootScope.goToState = function(state, params){
-        $state.go(state, params);
+    $rootScope.goToState = function(stateName, stateParameters){
+        if(stateName.indexOf('button') !== -1){
+            var buttonName = stateName;
+            stateName = $rootScope.appSettings.appDesign.floatingActionButton.active[buttonName].stateName;
+            stateParameters = $rootScope.appSettings.appDesign.floatingActionButton.active[buttonName].stateParameters;
+        }
+        $state.go(stateName, stateParameters);
     };
     $ionicPlatform.registerBackButtonAction(function (event) {
         if($rootScope.backButtonState){
@@ -201,10 +204,10 @@ angular.module('starter',
         }
         if($ionicHistory.currentStateName() === 'app.upgrade'){
             console.debug('registerBackButtonAction from upgrade: Going to default state...');
-            $state.go(config.appSettings.defaultState);
+            $state.go(config.appSettings.appDesign.defaultState);
             return;
         }
-        if($ionicHistory.currentStateName() === config.appSettings.defaultState){
+        if($ionicHistory.currentStateName() === config.appSettings.appDesign.defaultState){
             ionic.Platform.exitApp();
             return;
         }
@@ -215,7 +218,7 @@ angular.module('starter',
         if(localStorage.user){
             $rootScope.hideNavigationMenu = false;
             console.debug('registerBackButtonAction: Going to default state...');
-            $state.go(config.appSettings.defaultState);
+            $state.go(config.appSettings.appDesign.defaultState);
             return;
         }
         console.debug('registerBackButtonAction: Closing the app');
@@ -223,27 +226,15 @@ angular.module('starter',
     }, 100);
 
     var intervalChecker = setInterval(function(){if(typeof config !== "undefined"){clearInterval(intervalChecker);}}, 500);
-    function setIntoSeenAndOnboarded() {
-        var urlParameters = {};
-        var queryString = document.location.toString().split('?')[1];
-        var sURLVariables, parameterNameValueArray;
-        if (queryString) {sURLVariables = queryString.split('&');}
-        if (sURLVariables) {
-            for (var i = 0; i < sURLVariables.length; i++) {
-                parameterNameValueArray = sURLVariables[i].split('=');
-                if (parameterNameValueArray[1].indexOf('http') > -1) {urlParameters[parameterNameValueArray[0].toCamel()] = parameterNameValueArray[1];} else {urlParameters[parameterNameValueArray[0].toCamel()] = decodeURIComponent(parameterNameValueArray[1]);}
-            }
-        }
-        if (urlParameters.existingUser || urlParameters.introSeen || urlParameters.refreshUser) {
-            window.localStorage.introSeen = true;
-            window.localStorage.onboarded = true;
-        }
+    if (appsManager.getUrlParameter('existingUser') || appsManager.getUrlParameter('introSeen') || appsManager.getUrlParameter('refreshUser')) {
+        window.localStorage.introSeen = true;
+        window.localStorage.onboarded = true;
     }
-    setIntoSeenAndOnboarded();
 })
 .config(function($stateProvider, $urlRouterProvider, $compileProvider, ionicTimePickerProvider, ionicDatePickerProvider, $ionicConfigProvider, AnalyticsProvider) {
+    if(appsManager.getUrlParameter('debug')){window.debugMode = true;}
     if(appsManager.getUrlParameter('apiUrl')){localStorage.setItem('apiUrl', "https://" + appsManager.getUrlParameter('apiUrl'));}
-    var analyticsOptions = {tracker: 'UA-39222734-25', trackEvent: true};
+    var analyticsOptions = {tracker: 'UA-98995949-4', trackEvent: true};
     if(ionic.Platform.isAndroid()){
         var clientId = window.localStorage.GA_LOCAL_STORAGE_KEY;
         if(!clientId){
@@ -277,41 +268,16 @@ angular.module('starter',
         var i = this.length;
         while (i--) {if (this[i] === obj) {return true;}}
     };
-    var config_resolver = {};
-    window.config = {};
-    appsManager.getAppSettingsFromUrlParameter();
-    function getLocalJsonFile(path, successHandler) {return function($http){$http.get(path).success(successHandler);};}
-    function getLocalConfigJson(clientId) {return getLocalJsonFile('configs/' + clientId + '.config.json', function(response) {window.config.appSettings = response;});}
-    function getLocalPrivateConfigJson(clientId) {return getLocalJsonFile('private_configs/' + clientId + '.private_config.json', function(response) {window.private_keys = response;});}
-    if(!window.config.appSettings){
-        if(appsManager.doWeHaveLocalConfigFile()) {
-            config_resolver.appSettingsResponse = getLocalConfigJson(appsManager.getQuantiModoClientId());
-            config_resolver.privateKeysResponse = getLocalPrivateConfigJson(appsManager.getQuantiModoClientId());
-            //config_resolver.loadMyService = ['$ocLazyLoad', function($ocLazyLoad) {return $ocLazyLoad.load([appsManager.getAppConfig(), appsManager.getPrivateConfig()]);}];
-        } else {
-            var localStorageName = appsManager.getQuantiModoClientId() + 'AppSettings';
-            var locallyStoredAppSettings = localStorage.getItem(localStorageName);
-            if(!appsManager.getUrlParameter('refreshAppSettings') && locallyStoredAppSettings) {
-                window.config.appSettings = JSON.parse(locallyStoredAppSettings);
-            } else {
-                config_resolver.appSettingsResponse = function ($http) {
-                    return $http.get('https://app.quantimo.do/api/v1/appSettings?clientId=' + appsManager.getQuantiModoClientId()).then(function (response) {
-                        localStorage.setItem(localStorageName, JSON.stringify(response.data.data));
-                        window.config.appSettings = response.data.data;
-                    }, function errorCallback(response) {
-                        return getLocalConfigJson('quantimodo');
-                    });
-                };
-            }
-        }
-    }
-    var getOnboardingPages = ['$http', function($http) {return $http({method: 'GET', url: 'data/onboardingPages.json'});}];
-    var intro_resolver = config_resolver;
-    intro_resolver.introSlides = ['$http', function($http) {return $http({method: 'GET', url: 'data/introSlides.json'});}];
-    intro_resolver.onboardingPages = getOnboardingPages;
-    var onboarding_resolver = {onboardingPages: getOnboardingPages};
-    var inbox_resolver = {helpCards: ['$http', function($http) {return $http({method: 'GET', url: 'data/helpCards.json'});}]};
 
+    var config_resolver = {
+        appSettingsResponse: function($http){
+            var settingsUrl = 'configs/default.config.json';
+            var clientId = appsManager.getClientIdFromQueryParameters(true);
+            if(!appsManager.shouldWeUseLocalConfig()){settingsUrl = appsManager.getQuantiModoApiUrl() + '/api/v1/appSettings?clientId=' + clientId;}
+            return $http({method: 'GET', url: settingsUrl});
+        }
+    };
+    //config_resolver.loadMyService = ['$ocLazyLoad', function($ocLazyLoad) {return $ocLazyLoad.load([appsManager.getAppConfig(), appsManager.getPrivateConfig()]);}];
     ionicTimePickerProvider.configTimePicker({format: 12, step: 1, closeLabel: 'Cancel'});
     var datePickerObj = {
         inputDate: new Date(),
@@ -336,7 +302,7 @@ angular.module('starter',
             url: '/',
             templateUrl: 'templates/intro-tour-new.html',
             controller: 'IntroCtrl',
-            resolve : intro_resolver
+            resolve : config_resolver
         })
         .state('app', {
             url: "/app",
@@ -378,7 +344,8 @@ angular.module('starter',
                     templateUrl: "templates/intro-tour-new.html",
                     controller: 'IntroCtrl'
                 }
-            }
+            },
+            resolve : config_resolver
         })
         .state('app.track', {
             url: "/track",
@@ -1197,8 +1164,7 @@ angular.module('starter',
                     templateUrl: "templates/reminders-inbox.html",
                     controller: 'RemindersInboxCtrl'
                 }
-            },
-            resolve : inbox_resolver
+            }
         })
         .state('app.remindersInboxCompact', {
             url: "/reminders-inbox-compact",
@@ -1219,8 +1185,7 @@ angular.module('starter',
                     templateUrl: "templates/reminders-inbox.html",
                     controller: 'RemindersInboxCtrl'
                 }
-            },
-            resolve : inbox_resolver
+            }
         })
         .state('app.favorites', {
             url: "/favorites",
@@ -1245,7 +1210,7 @@ angular.module('starter',
             url: "/configuration",
             views: {
                 'menuContent': {
-                    templateUrl: "configuration/configuration.html",
+                    templateUrl: "../../app-configuration/configuration.html",
                     controller: 'ConfigurationCtrl'
                 }
             }
@@ -1278,8 +1243,7 @@ angular.module('starter',
                     templateUrl: "templates/reminders-inbox.html",
                     controller: 'RemindersInboxCtrl'
                 }
-            },
-            resolve : inbox_resolver
+            }
         })
         .state('app.remindersInboxTodayCategory', {
             url: "/reminders-inbox-today/:variableCategoryName",
@@ -1296,8 +1260,7 @@ angular.module('starter',
                     templateUrl: "templates/reminders-inbox.html",
                     controller: 'RemindersInboxCtrl'
                 }
-            },
-            resolve : inbox_resolver
+            }
         })
         .state('app.manageScheduledMeds', {
             url: "/manage-scheduled-meds",
@@ -1327,8 +1290,7 @@ angular.module('starter',
                     templateUrl: "templates/reminders-inbox.html",
                     controller: 'RemindersInboxCtrl'
                 }
-            },
-            resolve : inbox_resolver
+            }
         })
         .state('app.asNeededMeds', {
             url: "/as-needed-meds",
@@ -1357,8 +1319,7 @@ angular.module('starter',
                     templateUrl: "templates/reminders-inbox.html",
                     controller: 'RemindersInboxCtrl'
                 }
-            },
-            resolve : inbox_resolver
+            }
         })
         .state('app.remindersManage', {
             cache: false,
@@ -1409,8 +1370,7 @@ angular.module('starter',
                     templateUrl: "templates/onboarding-page.html",
                     controller: 'OnboardingCtrl'
                 }
-            },
-            resolve : onboarding_resolver
+            }
         })
         .state('app.upgrade', {
             url: "/upgrade",
